Relocate parser directory.
