+++ dojo/tools/awssecurityhub/parser.py
